extract filter button from stroll button
